Move component props to same file
