+++ src/components/molecules/DashboardItemSmall.tsx
@@ -1,8 +1,14 @@
 import React from 'react';
-import { DashboardItemSmall as DashboardItemSmallT } from '../../types/DashboardItem';
 import { Card, Pane, Heading } from 'evergreen-ui';
 
-const DashboardItemSmall: React.FC<DashboardItemSmallT> = ({ title, children, width, height, titleSize }) => {
+export type DashboardItemSmallProps = {
+    title: string;
+    titleSize: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
+    width: number | string;
+    height: number | string;
+};
+
+const DashboardItemSmall: React.FC<DashboardItemSmallProps> = ({ title, children, width, height, titleSize }) => {
     return (
         <Card width={width} height={height} display="flex" flexDirection="column" border="default" elevation={1}>
             <Heading size={titleSize} marginTop={25} marginLeft={25}>
